Move getModuleLink to utils
